# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for the Bookmark Management Service.",
        "title": "Bookmark Management API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/gen-pass": {
            "get": {
                "description": "Generate a random password with specified criteria",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Password"
                ],
                "summary": "Generate a random password",
                "responses": {
                    "200": {
                        "description": "Generated password",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to generate password",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/link/redirect/{code}": {
            "get": {
                "description": "Retrieves the original URL associated with the given code. Browser clients are redirected (301); API/JSON clients receive the URL in the response body.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ShortenURL"
                ],
                "summary": "Redirect to original URL by code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shortened URL code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns original URL (API/JSON client)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "301": {
                        "description": "Redirects to the original URL (browser)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Code is required or not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/link/shorten": {
            "post": {
                "description": "Create a short key for the provided URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ShortenURL"
                ],
                "summary": "Shorten a URL",
                "parameters": [
                    {
                        "description": "URL to shorten",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.shortenURLRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Shortened key",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.shortenURLRequest": {
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        }
    }
}